Separate functions into multiple files in the src directory
